requirements.txt
# Python Dependencies for Ultimate Email Validation System

Flask==2.0.3
requests==2.26.0
dnspython==2.1.0
redis==4.1.0
pytest==6.2.5
celery==5.2.3
SQLAlchemy==1.4.32
psycopg2-binary==2.9.3